Move plugin base classes to custodia.plugin
