+++ custodia/store/interface.py
@@ -1,37 +1,10 @@
 # Copyright (C) 2015  Custodia Project Contributors - see LICENSE file
+import warnings
 
-import logging
+from custodia.plugin import CSStore, CSStoreError, CSStoreExists
 
-from custodia.log import CustodiaPlugin
-
-logger = logging.getLogger(__name__)
-
-
-class CSStoreError(Exception):
-    def __init__(self, message=None):
-        logger.debug(message)
-        super(CSStoreError, self).__init__(message)
+__all__ = ('CSStore', 'CSStoreError', 'CSStoreExists')
 
 
-class CSStoreExists(Exception):
-    def __init__(self, message=None):
-        logger.debug(message)
-        super(CSStoreExists, self).__init__(message)
-
-
-class CSStore(CustodiaPlugin):
-
-    def get(self, key):
-        raise NotImplementedError
-
-    def set(self, key, value, replace=False):
-        raise NotImplementedError
-
-    def span(self, key):
-        raise NotImplementedError
-
-    def list(self, keyfilter=None):
-        raise NotImplementedError
-
-    def cut(self, key):
-        raise NotImplementedError
+warnings.warn('custodia.store.interface is deprecated, import from '
+              'custodia.plugin instead.', DeprecationWarning,)
