+++ tests/conftest.py
@@ -3,4 +3,7 @@ import warnings
 
 from custodia.log import ProvisionalWarning
 
+# silence our own warnings about provisional APIs
 warnings.simplefilter('ignore', category=ProvisionalWarning)
+# deprecated APIs raise an exception
+warnings.simplefilter('error', category=DeprecationWarning)
